requirements.txt
python-telegram-bot
google-genai
pymongo
python-dotenv
flask
dnspython
